@@ -23,10 +23,11 @@
     "issues": "https://github.com/phar-io/manifest/issues"
   },
   "require": {
-    "php": "^5.6 || ^7.0",
+    "php": "^7.2 || ^8.0",
     "ext-dom": "*",
     "ext-phar": "*",
-    "phar-io/version": "^2.0"
+    "ext-xmlwriter": "*",
+    "phar-io/version": "^3.0.1"
   },
   "autoload": {
     "classmap": [
@@ -35,8 +36,7 @@
   },
   "extra": {
     "branch-alias": {
-        "dev-master": "1.0.x-dev"
+        "dev-master": "2.0.x-dev"
     }
   }
 }
-
