@@ -21,45 +21,42 @@
     },
     "prefer-stable": true,
     "require": {
-        "php": "^7.1",
+        "php": ">=7.3",
         "ext-dom": "*",
         "ext-json": "*",
         "ext-libxml": "*",
         "ext-mbstring": "*",
         "ext-xml": "*",
-        "doctrine/instantiator": "^1.1",
-        "myclabs/deep-copy": "^1.7",
-        "phar-io/manifest": "^1.0.2",
-        "phar-io/version": "^2.0",
-        "phpspec/prophecy": "^1.7",
-        "phpunit/php-code-coverage": "^6.0.7",
-        "phpunit/php-file-iterator": "^2.0.1",
-        "phpunit/php-text-template": "^1.2.1",
-        "phpunit/php-timer": "^2.1",
-        "sebastian/comparator": "^3.0",
-        "sebastian/diff": "^3.0",
-        "sebastian/environment": "^4.0",
-        "sebastian/exporter": "^3.1",
-        "sebastian/global-state": "^2.0",
-        "sebastian/object-enumerator": "^3.0.3",
-        "sebastian/resource-operations": "^2.0",
-        "sebastian/version": "^2.0.1"
-    },
-    "require-dev": {
-        "ext-PDO": "*"
-    },
-    "conflict": {
-        "phpunit/phpunit-mock-objects": "*"
+        "ext-xmlwriter": "*",
+        "doctrine/instantiator": "^1.3.1",
+        "myclabs/deep-copy": "^1.10.1",
+        "phar-io/manifest": "^2.0.3",
+        "phar-io/version": "^3.0.2",
+        "phpunit/php-code-coverage": "^9.2.13",
+        "phpunit/php-file-iterator": "^3.0.5",
+        "phpunit/php-invoker": "^3.1.1",
+        "phpunit/php-text-template": "^2.0.3",
+        "phpunit/php-timer": "^5.0.2",
+        "sebastian/cli-parser": "^1.0.1",
+        "sebastian/code-unit": "^1.0.6",
+        "sebastian/comparator": "^4.0.5",
+        "sebastian/diff": "^4.0.3",
+        "sebastian/environment": "^5.1.3",
+        "sebastian/exporter": "^4.0.3",
+        "sebastian/global-state": "^5.0.1",
+        "sebastian/object-enumerator": "^4.0.3",
+        "sebastian/resource-operations": "^3.0.3",
+        "sebastian/type": "^3.1",
+        "sebastian/version": "^3.0.2"
     },
     "config": {
         "platform": {
-            "php": "7.1.0"
+            "php": "7.3.0"
         },
         "optimize-autoloader": true,
         "sort-packages": true
     },
     "suggest": {
-        "phpunit/php-invoker": "^2.0",
         "ext-soap": "*",
         "ext-xdebug": "*"
     },
@@ -69,6 +66,9 @@
     "autoload": {
         "classmap": [
             "src/"
+        ],
+        "files": [
+            "src/Framework/Assert/Functions.php"
         ]
     },
     "autoload-dev": {
@@ -76,7 +76,6 @@
             "tests/"
         ],
         "files": [
-            "src/Framework/Assert/Functions.php",
             "tests/_files/CoverageNamespacedFunctionTest.php",
             "tests/_files/CoveredFunction.php",
             "tests/_files/NamespaceCoveredFunction.php"
@@ -84,7 +83,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "7.5-dev"
+            "dev-master": "9.5-dev"
         }
     }
 }
