@@ -15,14 +15,14 @@ class GetNumberingRange extends Template implements CreateTemplate
      *
      * @var string
      */
-    public $Action = 'http://wcf.dian.colombia/IWcfDianCustomerServices/GetNumberingRange';
+    public string $Action = 'http://wcf.dian.colombia/IWcfDianCustomerServices/GetNumberingRange';
 
     /**
      * Required properties.
      *
      * @var array
      */
-    protected $requiredProperties = [
+    protected array $requiredProperties = [
         'Nit',
 		'IDSoftware',
     ];
@@ -31,11 +31,11 @@ class GetNumberingRange extends Template implements CreateTemplate
      * Construct.
      *
      * @param string $pathCertificate
-     * @param string $passwors
+     * @param string $password
      */
-    public function __construct($pathCertificate, $passwors)
+    public function __construct($pathCertificate, $password)
     {
-        parent::__construct($pathCertificate, $passwors);
+        parent::__construct($pathCertificate, $password);
         $this->To = 'https://vpfe.dian.gov.co/WcfDianCustomerServices.svc?wsdl';
     }
 
@@ -44,7 +44,7 @@ class GetNumberingRange extends Template implements CreateTemplate
      *
      * @return string
      */
-    public function createTemplate()
+    public function createTemplate(): string
     {
         return $this->templateXMLSOAP = <<<XML
 <soap:Envelope xmlns:soap="http://www.w3.org/2003/05/soap-envelope" xmlns:wcf="http://wcf.dian.colombia">
