@@ -15,24 +15,24 @@ class SendEvent extends Template implements CreateTemplate
      *
      * @var string
      */
-    public $Action = 'http://wcf.dian.colombia/IWcfDianCustomerServices/SendEventUpdateStatus';
+    public string $Action = 'http://wcf.dian.colombia/IWcfDianCustomerServices/SendEventUpdateStatus';
 
     /**
      * Required properties.
      *
      * @var array
      */
-    protected $requiredProperties = ['contentFile'];
+    protected array $requiredProperties = ['contentFile'];
 
     /**
      * Construct.
      *
      * @param string $pathCertificate
-     * @param string $passwors
+     * @param string $password
      */
-    public function __construct($pathCertificate, $passwors)
+    public function __construct($pathCertificate, $password)
     {
-        parent::__construct($pathCertificate, $passwors);
+        parent::__construct($pathCertificate, $password);
         $this->To = 'https://vpfe.dian.gov.co/WcfDianCustomerServices.svc?wsdl';
     }
 
@@ -41,7 +41,7 @@ class SendEvent extends Template implements CreateTemplate
      *
      * @return string
      */
-    public function createTemplate()
+    public function createTemplate(): string
     {
         return $this->templateXMLSOAP = <<<XML
 <soap:Envelope xmlns:soap="http://www.w3.org/2003/05/soap-envelope" xmlns:wcf="http://wcf.dian.colombia">
