@@ -1,8 +1,9 @@
 <?php
 
-namespace ubl21dian\Tests;
+namespace ubl21dian\test;
 
 use DOMDocument;
+use ubl21dian\test\TestCase;
 use ubl21dian\XAdES\SignInvoice;
 
 /**
@@ -603,9 +604,9 @@ XML;
     public function it_generates_signature_XAdES_sha1()
     {
         $pathCertificate = dirname(dirname(__FILE__)).'/certicamara.p12';
-        $passwors = '3T3rN4661343';
+        $password = '3T3rN4661343';
 
-        $signInvoice = new SignInvoice($pathCertificate, $passwors, $this->xmlString, SignInvoice::ALGO_SHA1);
+        $signInvoice = new SignInvoice($pathCertificate, $password, $this->xmlString, SignInvoice::ALGO_SHA1);
 
         $domDocumentValidate = new DOMDocument();
         $domDocumentValidate->validateOnParse = true;
@@ -619,9 +620,9 @@ XML;
     public function it_generates_signature_XAdES_sha256()
     {
         $pathCertificate = dirname(dirname(__FILE__)).'/certicamara.p12';
-        $passwors = '3T3rN4661343';
+        $password = '3T3rN4661343';
 
-        $signInvoice = new SignInvoice($pathCertificate, $passwors, $this->xmlString);
+        $signInvoice = new SignInvoice($pathCertificate, $password, $this->xmlString);
 
         $domDocumentValidate = new DOMDocument();
         $domDocumentValidate->validateOnParse = true;
@@ -635,9 +636,9 @@ XML;
     public function it_generates_signature_XAdES_sha512()
     {
         $pathCertificate = dirname(dirname(__FILE__)).'/certicamara.p12';
-        $passwors = '3T3rN4661343';
+        $password = '3T3rN4661343';
 
-        $signInvoice = new SignInvoice($pathCertificate, $passwors, $this->xmlString, SignInvoice::ALGO_SHA512);
+        $signInvoice = new SignInvoice($pathCertificate, $password, $this->xmlString, SignInvoice::ALGO_SHA512);
 
         $domDocumentValidate = new DOMDocument();
         $domDocumentValidate->validateOnParse = true;
@@ -651,9 +652,9 @@ XML;
     public function it_generates_signature_XAdES_and_software_security_code()
     {
         $pathCertificate = dirname(dirname(__FILE__)).'/certicamara.p12';
-        $passwors = '3T3rN4661343';
+        $password = '3T3rN4661343';
 
-        $signInvoice = new SignInvoice($pathCertificate, $passwors);
+        $signInvoice = new SignInvoice($pathCertificate, $password);
 
         // Software security code
         $signInvoice->softwareID = 'xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx';
@@ -673,9 +674,9 @@ XML;
     public function it_generates_signature_XAdES_and_calculate_cufe()
     {
         $pathCertificate = dirname(dirname(__FILE__)).'/certicamara.p12';
-        $passwors = '3T3rN4661343';
+        $password = '3T3rN4661343';
 
-        $signInvoice = new SignInvoice($pathCertificate, $passwors);
+        $signInvoice = new SignInvoice($pathCertificate, $password);
 
         // CUFE
         $signInvoice->technicalKey = 'xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx';
