@@ -1,8 +1,9 @@
 <?php
 
-namespace ubl21dian\Tests;
+namespace ubl21dian\test;
 
 use DOMDocument;
+use ubl21dian\test\TestCase;
 use ubl21dian\XAdES\SignCreditNote;
 
 /**
@@ -201,9 +202,9 @@ XML;
     public function it_generates_signature_XAdES_sha1()
     {
         $pathCertificate = dirname(dirname(__FILE__)).'/certicamara.p12';
-        $passwors = '3T3rN4661343';
+        $password = '3T3rN4661343';
 
-        $signCreditNote = new SignCreditNote($pathCertificate, $passwors, $this->xmlString, SignCreditNote::ALGO_SHA1);
+        $signCreditNote = new SignCreditNote($pathCertificate, $password, $this->xmlString, SignCreditNote::ALGO_SHA1);
 
         $domDocumentValidate = new DOMDocument();
         $domDocumentValidate->validateOnParse = true;
@@ -217,9 +218,9 @@ XML;
     public function it_generates_signature_XAdES_sha256()
     {
         $pathCertificate = dirname(dirname(__FILE__)).'/certicamara.p12';
-        $passwors = '3T3rN4661343';
+        $password = '3T3rN4661343';
 
-        $signCreditNote = new SignCreditNote($pathCertificate, $passwors, $this->xmlString);
+        $signCreditNote = new SignCreditNote($pathCertificate, $password, $this->xmlString);
 
         $domDocumentValidate = new DOMDocument();
         $domDocumentValidate->validateOnParse = true;
@@ -233,9 +234,9 @@ XML;
     public function it_generates_signature_XAdES_sha512()
     {
         $pathCertificate = dirname(dirname(__FILE__)).'/certicamara.p12';
-        $passwors = '3T3rN4661343';
+        $password = '3T3rN4661343';
 
-        $signCreditNote = new SignCreditNote($pathCertificate, $passwors, $this->xmlString, SignCreditNote::ALGO_SHA512);
+        $signCreditNote = new SignCreditNote($pathCertificate, $password, $this->xmlString, SignCreditNote::ALGO_SHA512);
 
         $domDocumentValidate = new DOMDocument();
         $domDocumentValidate->validateOnParse = true;
@@ -249,9 +250,9 @@ XML;
     public function it_generates_signature_XAdES_and_software_security_code()
     {
         $pathCertificate = dirname(dirname(__FILE__)).'/certicamara.p12';
-        $passwors = '3T3rN4661343';
+        $password = '3T3rN4661343';
 
-        $signCreditNote = new SignCreditNote($pathCertificate, $passwors);
+        $signCreditNote = new SignCreditNote($pathCertificate, $password);
 
         // Software security code
         $signCreditNote->softwareID = 'xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx';
@@ -271,9 +272,9 @@ XML;
     public function it_generates_signature_XAdES_and_calculate_cude()
     {
         $pathCertificate = dirname(dirname(__FILE__)).'/certicamara.p12';
-        $passwors = '3T3rN4661343';
+        $password = '3T3rN4661343';
 
-        $signCreditNote = new SignCreditNote($pathCertificate, $passwors);
+        $signCreditNote = new SignCreditNote($pathCertificate, $password);
 
         // CUDE
         $signCreditNote->pin = 'xxxxx';
